==> estatistica-python/01-estruturacao-dados.ipynb <==
# Projeto Análise Estatística do ENEM 2019

Este projeto analisa os resultados do ENEM no estado de São Paulo no ano de 2019.

Os dados foram extraídos do site do INEP:

inep.gov.br/microdados

# 01-Preparação, organização e estruturação dos dados

import numpy as np
import pandas as pd

dados = pd.read_csv('./microdados_enem_2019_sp.csv', sep=';', encoding='iso-8859-1')

dados.head()

dados.shape

## Extração e Limpeza dos Dados

# Análise dos tipos de atributos

dados.dtypes

dados1 = dados.drop(columns=['CO_MUNICIPIO_RESIDENCIA'])

dados1.head()

dados1 = dados1.drop(columns=['CO_UF_RESIDENCIA', 'SG_UF_RESIDENCIA', 'CO_MUNICIPIO_NASCIMENTO'])

dados1 = dados1.drop(columns=['NO_MUNICIPIO_NASCIMENTO', 'CO_UF_NASCIMENTO', 'SG_UF_NASCIMENTO',
                             'TP_ANO_CONCLUIU', 'TP_ENSINO', 'CO_MUNICIPIO_ESC', 'CO_UF_ESC',
                             'SG_UF_ESC'])

dados1 = dados1.drop(columns=['TP_DEPENDENCIA_ADM_ESC', 'TP_LOCALIZACAO_ESC', 'TP_SIT_FUNC_ESC'])

dados1.shape

### Corrigindo erros nas notas

# Todas as notas estão com um erro de uma casa decimal

dados1.loc[:, 'NU_NOTA_CN'] /= 10 #  /=  dividir todos os elementos da coluna

dados1.head()

dados1.loc[:, 'NU_NOTA_CH'] /= 10
dados1.loc[:, 'NU_NOTA_LC'] /= 10
dados1.loc[:, 'NU_NOTA_MT'] /= 10

dados1.head()

### Renomeando Variáveis (colunas)

dados1 = dados1.rename(columns={'NU_NOTA_REDACAO': 'NOTA_REDACAO'})

dados1.head()

dados1 = dados1.rename(columns={'NU_NOTA_CN': 'NOTA_CN', 'NU_NOTA_CH': 'NOTA_CH', 
                                'NU_NOTA_LC': 'NOTA_LC', 'NU_NOTA_MT': 'NOTA_MT' })

dados1 = dados1.rename(columns={'NU_NOTA_COMP1': 'COMP1', 'NU_NOTA_COMP2': 'COMP2', 
                                'NU_NOTA_COMP3': 'COMP3', 'NU_NOTA_COMP4': 'COMP4',
                                'NU_NOTA_COMP5': 'COMP5'})

dados1 = dados1.rename(columns={'NU_IDADE': 'IDADE', 'TP_SEXO': 'SEXO', 
                                'TP_COR_RACA': 'RACA', 'Q025': 'INTERNET',
                                'TP_ESCOLA': 'ESCOLA'})

### Renomeando Registros (Linhas)

dados1['RACA'] = dados1['RACA'].replace({0:'nao_declarado', 1:'branca', 2:'preta', 3:'parda', 4:'amarela', 5:'indigena'})

dados1.head()

dados1['TP_LINGUA'] = dados1['TP_LINGUA'].replace({0:'Ingles', 1:'Espanhol'})

dados1['ESCOLA'] = dados1['ESCOLA'].replace({1:'nao_respondeu', 2:'publica', 3:'privada', 4:'exterior'})

dados1['INTERNET'] = dados1['INTERNET'].replace({'A':'Nao', 'B':'Sim'})

## Análise Exploratória e Transformação dos Dados

### Análise das idades

dados1['IDADE'].value_counts()

dados1['IDADE'].value_counts().sort_index()

# Alguns valores estranhos e curiosos, como idades menores que 12 anos e maior de 80 anos

menores_12 = dados1.query('IDADE < 12')['NO_MUNICIPIO_RESIDENCIA'].value_counts()

menores_12

# removendo os dados com idades menores de 12 anos
dados_maiores_11 = dados1.loc[dados1.IDADE > 11]

dados_maiores_11.head()

dados_maiores_11['IDADE'].value_counts().sort_index()

### Treineiros

dados_maiores_11.query('IN_TREINEIRO == 1')['IN_TREINEIRO'].value_counts()

treineiros = dados_maiores_11.loc[dados_maiores_11.IN_TREINEIRO == 1]

treineiros.head()

treineiros.to_csv('treineiros_enem_2019_sp.csv', encoding = 'iso-8859-1', index = False)

# Separação entre os treineiros e os vestibulandos
vestibulandos = dados_maiores_11.loc[dados_maiores_11.IN_TREINEIRO == 0]

vestibulandos.head()

vestibulandos.shape

### Análise das presenças

vestibulandos['TP_PRESENCA_CN'].value_counts()

vestibulandos['TP_PRESENCA_CH'].value_counts()

vestibulandos['TP_PRESENCA_LC'].value_counts()

vestibulandos['TP_PRESENCA_MT'].value_counts()

vestibulandos['TP_STATUS_REDACAO'].value_counts().sort_index()

### Valores Ausentes

vestibulandos['IDADE'].isnull().sum()

provas = ['NOTA_REDACAO', 'NOTA_CN', 'NOTA_CH', 'NOTA_LC', 'NOTA_MT']

# Registros NAN nas notas
vestibulandos[provas].isnull().sum()

vestibulandos_classificados = vestibulandos.loc[vestibulandos.TP_PRESENCA_CH == 1]
vestibulandos_classificados = vestibulandos_classificados.loc[vestibulandos_classificados.TP_PRESENCA_MT == 1]

vestibulandos_classificados['TP_PRESENCA_MT'].value_counts()

vestibulandos_classificados['TP_PRESENCA_CH'].value_counts()

vestibulandos_classificados['TP_PRESENCA_CN'].value_counts()

vestibulandos_classificados['TP_PRESENCA_LC'].value_counts()

vestibulandos_classificados[provas].isnull().sum()

### Notas zero na redação

vestibulandos_classificados.query('NOTA_REDACAO == 0')['NOTA_REDACAO'].value_counts()

vestibulandos_classificados.query('COMP1 == 0')['COMP1'].value_counts()

vestibulandos_classificados.query('COMP2 == 0')['COMP2'].value_counts()

vestibulandos_classificados.query('COMP3 == 0')['COMP3'].value_counts()

vestibulandos_classificados.query('COMP4 == 0')['COMP4'].value_counts()

vestibulandos_classificados.query('COMP5 == 0')['COMP5'].value_counts()

# Componente 5 é um ponto fraco da redação

vestibulandos_classificados.query('NOTA_CH == 0')['NOTA_CH'].value_counts()

vestibulandos_classificados.query('NOTA_CN == 0')['NOTA_CN'].value_counts()

vestibulandos_classificados.query('NOTA_LC == 0')['NOTA_LC'].value_counts()

vestibulandos_classificados.query('NOTA_MT == 0')['NOTA_MT'].value_counts()

vestibulandos_classificados.head()

vestibulandos_classificados = vestibulandos_classificados.drop(columns=['IN_TREINEIRO'])

vestibulandos_classificados.shape

## Salvando (Exportando) o Dataframe tratado

vestibulandos_classificados.to_csv('enem_2019_tratado.csv', encoding = 'iso-8859-1', index = False)
