==> estatistica-python/02-fundamentos-estatistica.ipynb <==
# 02-Fundamentos da Estatística

## Amostragens

import numpy as np
import pandas as pd

enem_sp = pd.read_csv('./enem_2019_tratado.csv', encoding = 'iso-8859-1', sep = ',')

enem_sp.head()

enem_sp.shape

enem_campinas = enem_sp.loc[enem_sp.NO_MUNICIPIO_RESIDENCIA == 'Campinas']

enem_campinas.head()

enem_campinas.shape

### Amostragem Aleatória Simples

np.random.seed(33)

amos_simples = enem_campinas.sample(frac = 0.2) # 20% dos dados

amos_simples

amos_simples = enem_campinas.sample(3000) # número fixo de dados

amos_simples

np.random.seed(1234)

amostra = np.random.choice(a = [0,1], size = 13198, replace = True, p = [0.8, 0.2])

amostra

len(amostra[amostra == 0])

len(amostra[amostra == 1])

amos_simples = enem_campinas.loc[amostra == 1]

amos_simples

### Amostragem Sistemática

enem_campinas.shape

np.random.seed(3124)
inicio = np.random.choice(10, 1)

amostra2 = np.arange(inicio, 13198, 100)
amostra2

amos_sistematica = enem_campinas.iloc[amostra2, :]

amos_sistematica

# outra maneira

np.random.seed(543)

inicio = np.random.choice(10, 1)
passo = np.random.randint(100, 150, 1)

amostra2 = np.arange(inicio, 13198, passo)

amostra2_camp = enem_campinas.iloc[amostra2, :]

amostra2_camp

### Amostragem Estratificada

enem_campinas['RACA'].value_counts()

# Para quantidade pequena de grupos

branca = enem_campinas.loc[enem_campinas.RACA == 'branca'].sample(frac = 0.1)
branca.shape

parda = enem_campinas.loc[enem_campinas.RACA == 'parda'].sample(frac = 0.1)
parda.shape

preta = enem_campinas.loc[enem_campinas.RACA == 'preta'].sample(frac = 0.1)
preta.shape

amarela = enem_campinas.loc[enem_campinas.RACA == 'amarela'].sample(frac = 0.1)
amarela.shape

nd = enem_campinas.loc[enem_campinas.RACA == 'nao_declarado'].sample(frac = 0.1)
nd.shape

indigena = enem_campinas.loc[enem_campinas.RACA == 'indigena'].sample(frac = 0.1)
indigena.shape

amos_estra = pd.concat([branca, parda, preta, amarela, nd, indigena])
amos_estra

# Para grandes quantidades de grupos

from sklearn.model_selection import train_test_split

x_treino, x_teste, y_treino, amos_estra = train_test_split(enem_campinas.drop('RACA', axis=1),
                                                          enem_campinas['RACA'], stratify=enem_campinas['RACA'],
                                                          test_size=0.2)

amos_estra.shape

amos_estra.value_counts()

amos_estra = amos_estra.index

amos_estra

amos_estra_camp = enem_campinas.loc[amos_estra, :]

amos_estra_camp

### Amostragem Por Conglomerado (Agrupamento)

escolas_publicas = enem_campinas.loc[enem_campinas.ESCOLA == 'publica']

escolas_publicas.shape

amostras = escolas_publicas['CO_ESCOLA'].value_counts()
amostras

amostras.index

amostragem = np.random.choice(a=[0,1], size=186,replace=True, p=[0.85, 0.15]) # 186 é a quantidade de escolas publicas (length)

amostragem

amos_conglomerado = amostras.loc[amostragem == 1]

amos_conglomerado

amos_conglomerado.shape

amos_conglomerado.index

amos_conglomerado = escolas_publicas.loc[escolas_publicas['CO_ESCOLA'].isin([35045664.0, 35018351.0, 35018314.0, 35018570.0, 35112859.0, 35018259.0,
       35905471.0, 35924544.0, 35812900.0, 35018296.0, 35467765.0, 35045676.0,
       35018922.0, 35065626.0, 35901118.0, 35045697.0, 35213640.0, 35112720.0,
       35131982.0, 35018200.0, 35165062.0, 35915592.0, 35112707.0, 35112641.0,
       35172510.0, 35112537.0, 35307701.0, 35917552.0, 35017255.0])]

amos_conglomerado

## Distribuição de Frequências

import numpy as np
import pandas as pd
from collections import Counter

enem_sp = pd.read_csv('./enem_2019_tratado.csv', sep=',', encoding='iso-8859-1')

enem_sp.head()

enem_sorocaba = enem_sp.loc[enem_sp.NO_MUNICIPIO_RESIDENCIA == 'Sorocaba']

enem_sorocaba.shape

### Frequência absoluta

idades = enem_sorocaba['IDADE']

idades

freq_absoluta = Counter(idades)

freq_absoluta

freq_absoluta = pd.DataFrame.from_dict(freq_absoluta, orient='index')

freq_absoluta

freq_absoluta = freq_absoluta.sort_index(ascending=True)

freq_absoluta

### Frequência Relativa

freq_relativa = freq_absoluta / freq_absoluta.sum()
freq_relativa

freq_relativa_perc = freq_relativa * 100
freq_relativa_perc

## Gráfico de frequências: Histograma

### Matplotlib

import matplotlib.pyplot as plt

enem_sorocaba.IDADE.hist(bins = 8)
plt.title('Distribuição de Idades - ENEM SP 2019 - Sorocaba')
plt.show()

plt.hist(enem_sorocaba.IDADE, 8, rwidth = 1)
plt.title('Distribuição de Idades - ENEM SP 2019 - Sorocaba')
plt.show()

from matplotlib import colors
plt.hist(enem_sorocaba.IDADE, 20, rwidth=1, color='green', density=True)
plt.title('Distribuição de Idades - ENEM SP 2019 - Sorocaba')
plt.xlabel('Idade')
plt.ylabel('Frequência')
plt.show()

### Seaborn

import seaborn as sns

sns.displot(enem_sorocaba['IDADE'], bins=8, color='blue', kde=False); # kde = linha de densidade

# podemos usar displot e histplot
sns.histplot(enem_sorocaba, x='IDADE', bins=8, color='green', kde=True, stat='probability');

### Plotly

import plotly.express as px
grafico = px.histogram(enem_sorocaba, x='IDADE', nbins=10, histnorm='percent')
grafico.update_layout(width=500,height=500,title_text='Distribuição de idades - ENEM SP 2019 - Sorocaba')
grafico.show()

grafico = px.histogram(enem_sorocaba, x='IDADE', nbins=10, histnorm='probability', color='SEXO')
grafico.update_layout(width=500,height=500,title_text='Distribuição de idades - ENEM SP 2019 - Sorocaba')
grafico.show()

grafico = px.histogram(enem_sorocaba, x='NOTA_MT', nbins=14, color='SEXO')
grafico.update_layout(width=500,height=500,title_text='Distribuição de idades - ENEM SP 2019 - Sorocaba')
grafico.show()

grafico = px.histogram(enem_sorocaba, x='NOTA_CN', nbins=14, color='SEXO')
grafico.update_layout(width=500,height=500,title_text='Distribuição de idades - ENEM SP 2019 - Sorocaba')
grafico.show()

grafico = px.histogram(enem_sorocaba, x='NOTA_REDACAO', nbins=20, color='RACA', histnorm='percent')
grafico.update_layout(width=500,height=500,title_text='Distribuição de idades - ENEM SP 2019 - Sorocaba')
grafico.show()

## Medidas de tendência central

import numpy as np 
import pandas as pd
enem_sp = pd.read_csv('./enem_2019_tratado.csv', sep=',', encoding='iso-8859-1')

### Média

enem_sp['IDADE'].mean()

round(enem_sp['IDADE'].mean(),1)

enem_sp['NOTA_REDACAO'].mean()

enem_sp.describe()

### Mediana

enem_sp['IDADE'].median()

### Moda

enem_sp.IDADE.mode()

### Distribuição

enem_sp.COMP3.mode()

enem_sp.COMP3.mean()

enem_sp.COMP3.median()

import plotly.express as px

grafico = px.histogram(enem_sp, x='COMP3', nbins=15)
grafico.update_layout(width=400, height=400, title_text='Nota do componente 3 da redação - ENEM SP 2019')
grafico.show()

enem_sp.COMP2.mode()

enem_sp.COMP2.mean()

enem_sp.COMP2.median()

grafico = px.histogram(enem_sp, x='COMP2', nbins=15)
grafico.update_layout(width=400, height=400, title_text='Nota do componente 2 da redação - ENEM SP 2019')
grafico.show()

## Medidas de Dispersão e Posição

import numpy as np 
import pandas as pd
enem_sp = pd.read_csv('./enem_2019_tratado.csv', sep=',', encoding='iso-8859-1')

### Medidas de Dispersão

#### Variância

enem_sp.NOTA_MT.var()

#### Desvio-padrão

enem_sp.NOTA_MT.std()

enem_sp.describe()

### Medidas de Posição

#### Mínimo

enem_sp.NOTA_MT.min()

#### Máximo

enem_sp.NOTA_MT.max()

#### Primeiro Quartil

enem_sp.NOTA_MT.quantile(0.25)

#### Segundo Quartil (mediana)

enem_sp.NOTA_MT.quantile(0.50)

#### Terceiro Quartil

enem_sp.NOTA_MT.quantile(0.75)

## Gráfico BoxPlot e Outliers

### Matplotlib

import matplotlib.pyplot as plt

plt.boxplot(enem_sp.NOTA_MT)
plt.show

### Seaborn

import seaborn as sns

sns.boxplot(y='NOTA_MT', data=enem_sp);

### Plotly

import plotly.express as px

grafico = px.box(enem_sp, y='NOTA_MT')
grafico.show()

sem_outliers = enem_sp.loc[enem_sp.NOTA_MT <= 902.9]

sem_outliers = sem_outliers.loc[sem_outliers.NOTA_MT >= 359.1]

grafico = px.box(sem_outliers, y='NOTA_MT')
grafico.show()

grafico = px.box(enem_sp, x='SEXO', y='NOTA_MT')
grafico.show()
